backend/app.py
# app.py
from flask import Flask, request, jsonify
import tensorflow as tf
from tensorflow.keras.preprocessing.image import load_img, img_to_array
from tensorflow.keras.applications import VGG16
from tensorflow.keras.applications.vgg16 import preprocess_input
import cv2
import numpy as np
from PIL import Image

app = Flask(__name__)

# Load pre-trained VGG16 model
model = VGG16(weights='imagenet', include_top=False, input_shape=(224, 224, 3))

food_names = {
    0: 'Broccoli',
    1: 'Carrots',
    2: 'Cauliflower',
    3: 'Kale',
    4: 'Mushrooms',
    5: 'Onions',
    6: 'Peppers',
    7: 'Potatoes',
    8: 'Spinach',
    9: 'Tomatoes'
}

@app.route('/recognize', methods=['POST'])
def recognize_food():
    image = request.files['image']
    img = load_img(image, target_size=(224, 224))
    img_array = img_to_array(img)
    img_array = preprocess_input(img_array)
    img_array = img_array.reshape((1, 224, 224, 3))

    # Make predictions using the pre-trained model
    predictions = model.predict(img_array)

    # Perform additional processing to extract features and make predictions
    # for dietary management, obesity control, health level balancing, food quality prediction, and density measuring
    # ...

    # Convert image to OpenCV format
    img_cv = cv2.imdecode(np.frombuffer(image.read(), np.uint8), cv2.IMREAD_COLOR)

    # Perform image processing and feature extraction using OpenCV
    # ...

    # Get the predicted food name
    predicted